+++ torch_frame/nn/base.py
@@ -14,7 +14,8 @@ class Module(torch.nn.Module):
 
     This is achieved by postponing submodule creation
     (via :meth:`init_modules`) until all attributes in :obj:`LAZY_ATTRS` are
-    fully-specified."""
+    fully-specified.
+    """
     LAZY_ATTRS: Set[str] = set()
 
     def init_modules(self):
@@ -67,9 +68,9 @@ class Module(torch.nn.Module):
         self.validate()
         return super().named_parameters(*args, **kwargs)
 
-    def named_children(self, *args, **kwargs) -> Any:
+    def named_children(self) -> Any:
         self.validate()
-        return super().named_children(*args, **kwargs)
+        return super().named_children()
 
     def named_modules(self, *args, **kwargs) -> Any:
         self.validate()