+++ torch_frame/nn/decoder/__init__.py
@@ -1,3 +1,4 @@
+r"""Decoder package."""
 from .decoder import Decoder
 from .trompt_decoder import TromptDecoder
 from .excelformer_decoder import ExcelFormerDecoder